git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13261 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/findfile.h
@@ -7,7 +7,7 @@
 
 // Empty file
 typedef struct FILEFINDSTRUCT {
-	unsigned long size;
+	uint32_t size;
 	char name [256];
 } FILEFINDSTRUCT;
